- Assignments now type-checked against function return types, including   cross-package checking

+++ test2.py
@@ -2,5 +2,5 @@ h: list = [1, 2, 3]
 
 a = "fdsa"
 
-def f(x, y) -> int:
+def f(x, y) -> float:
     return x + y
